requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.1
colorama==0.4.6
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.4.0
flake8==7.3.0
isort==6.0.1
mccabe==0.7.0
mypy==1.16.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
